// api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:8000' | (string & {});
};

/**
 * Entry
 */
export type Entry = {
    /**
     * Id
     */
    id?: number;
    /**
     * Title
     */
    title?: string;
    /**
     * Content
     */
    content?: string;
    /**
     * User Name
     */
    user_name?: string;
    /**
     * Created At
     */
    created_at?: string;
    /**
     * Updated At
     */
    updated_at?: string;
};

/**
 * EntryRequest
 */
export type EntryRequest = {
    /**
     * Title
     */
    title?: string;
    /**
     * Content
     */
    content?: string;
    /**
     * User Name
     */
    user_name?: string;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * MessageResponse
 */
export type MessageResponse = {
    /**
     * Message
     */
    message?: string;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type RootRootData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/';
};

export type RootRootResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type EntriesGetEntriesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/entries';
};

export type EntriesGetEntriesResponses = {
    /**
     * Response Entries Get Entries
     *
     * Successful Response
     */
    200: Array<Entry>;
};

export type EntriesGetEntriesResponse = EntriesGetEntriesResponses[keyof EntriesGetEntriesResponses];

export type EntriesCreateEntryData = {
    body: EntryRequest;
    path?: never;
    query?: never;
    url: '/entries';
};

export type EntriesCreateEntryErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type EntriesCreateEntryError = EntriesCreateEntryErrors[keyof EntriesCreateEntryErrors];

export type EntriesCreateEntryResponses = {
    /**
     * Successful Response
     */
    200: Entry;
};

export type EntriesCreateEntryResponse = EntriesCreateEntryResponses[keyof EntriesCreateEntryResponses];

export type EntriesDeleteEntryData = {
    body?: never;
    path: {
        /**
         * Entry Id
         */
        entry_id: number;
    };
    query?: never;
    url: '/entries/{entry_id}';
};

export type EntriesDeleteEntryErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type EntriesDeleteEntryError = EntriesDeleteEntryErrors[keyof EntriesDeleteEntryErrors];

export type EntriesDeleteEntryResponses = {
    /**
     * Successful Response
     */
    200: MessageResponse;
};

export type EntriesDeleteEntryResponse = EntriesDeleteEntryResponses[keyof EntriesDeleteEntryResponses];

export type EntriesUpdateEntryData = {
    body: EntryRequest;
    path: {
        /**
         * Entry Id
         */
        entry_id: number;
    };
    query?: never;
    url: '/entries/{entry_id}';
};

export type EntriesUpdateEntryErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type EntriesUpdateEntryError = EntriesUpdateEntryErrors[keyof EntriesUpdateEntryErrors];

export type EntriesUpdateEntryResponses = {
    /**
     * Successful Response
     */
    200: Entry;
};

export type EntriesUpdateEntryResponse = EntriesUpdateEntryResponses[keyof EntriesUpdateEntryResponses];